added check for moveable files

+++ wrf_run/post_processing/draw_plots.sh
@@ -2,7 +2,7 @@
 # @Author: Benjamin Held
 # @Date:   2017-03-12 16:04:54
 # @Last Modified by:   Benjamin Held
-# @Last Modified time: 2020-02-06 20:07:07
+# @Last Modified time: 2020-02-09 11:49:03
 
 # setting -e to abort on error
 set -e
@@ -45,10 +45,11 @@ mkdir "${DEST_FOLDER}/rain_3h"
 mkdir "${DEST_FOLDER}/rain_tot"
 mkdir "${DEST_FOLDER}/thunderstorm_index"
 
-mv comp_*.png "${DEST_FOLDER}/comp"
-mv rain_3h_*.png "${DEST_FOLDER}/rain_3h"
-mv rain_tot_*.png "${DEST_FOLDER}/rain_tot"
-mv thunderstorm_*.png "${DEST_FOLDER}/thunderstorm_index"
+# Check for moveable file and move them if present
+[ ! -f comp_*.png ] || mv comp_*.png "${DEST_FOLDER}/comp"
+[ ! -f rain_3h_*.png ] || mv rain_3h_*.png "${DEST_FOLDER}/rain_3h"
+[ ! -f rain_tot_*.png ] || mv rain_tot_*.png "${DEST_FOLDER}/rain_tot"
+[ ! -f thunderstorm_*.png ] || mv thunderstorm_*.png "${DEST_FOLDER}/thunderstorm_index"
 
 # generate meat.ini file
 cd "${SCRIPT_PATH}" || exit 1
